+++ GameSource/Areas/Admin/Controllers/UserController.cs
@@ -1,0 +1,239 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using GameSource.Areas.Admin.ViewModels.UserViewModel;
+using GameSource.Models.GameSourceUser;
+using GameSource.Models.GameSourceUser.Enums;
+using GameSource.Services.GameSourceUser.Contracts;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
+
+namespace GameSource.Areas.Admin.Controllers
+{
+    //[Authorize(Roles = "Admin")]
+    [Area("Admin")]
+    [Route("Admin/[controller]")]
+    public class UserController : Controller
+    {
+        private readonly IUserService userService;
+        private readonly IUserRoleService userRoleService;
+        private readonly IUserStatusService userStatusService;
+        private readonly UserManager<User> userManager;
+        private readonly SignInManager<User> signInManager;
+
+        public UserController(IUserService userService, IUserRoleService userRoleService, IUserStatusService userStatusService, UserManager<User> userManager, SignInManager<User> signInManager) 
+        {
+            this.userService = userService;
+            this.userRoleService = userRoleService;
+            this.userStatusService = userStatusService;
+            this.userManager = userManager;
+            this.signInManager = signInManager;
+        }
+
+        [HttpGet("Index")]
+        public async Task<IActionResult> Index()
+        {
+            AdminUserIndexViewModel viewModel = new AdminUserIndexViewModel
+            {
+                Users = await userService.GetAllAsync(),
+                UserRoles = await userRoleService.GetAllAsync(),
+                UserStatuses = await userStatusService.GetAllAsync()
+            };
+
+            return View("~/Areas/Admin/Views/User/Index.cshtml", viewModel);
+        }
+
+        [HttpGet("Details/{id}")]
+        public async Task<IActionResult> Details(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var user = await userService.GetByIDAsync((int)id);
+            if (user == null)
+            {
+                return NotFound();
+            }
+
+            AdminUserDetailsViewModel viewModel = new AdminUserDetailsViewModel
+            {
+                User = user,
+                UserRole = await userRoleService.GetByIDAsync(user.UserRoleID),
+                UserStatus = await userStatusService.GetByIDAsync(user.UserStatusID)
+            };
+
+            return View("~/Areas/Admin/Views/User/Details.cshtml", viewModel);
+        }
+
+        [HttpGet("Register")]
+        public IActionResult Register()
+        {
+            AdminUserRegisterViewModel viewModel = new AdminUserRegisterViewModel();
+            return View("~/Areas/Admin/Views/User/Register.cshtml", viewModel);
+        }
+
+        [HttpPost("Register")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Register(AdminUserRegisterViewModel viewModel)
+        {
+            if (ModelState.IsValid)
+            {
+                User user = new User
+                {
+                    UserName = viewModel.Username,
+                    Email = viewModel.Email,
+                    FirstName = viewModel.FirstName,
+                    LastName = viewModel.LastName,
+                    DateCreated = DateTime.Now,
+                    UserStatusID = (int)UserStatusEnum.Active,
+                    UserRoleID = (int)UserRoleEnum.Member
+                };
+
+                var result = await userManager.CreateAsync(user, viewModel.Password);
+                if (result.Succeeded)
+                {
+                    var newUser = await userManager.FindByNameAsync(user.UserName);
+                    var roleResult = await userManager.AddToRoleAsync(newUser, "Member");
+
+                    if (roleResult.Succeeded)
+                    {
+                        await signInManager.SignInAsync(user, isPersistent: false);
+                        return RedirectToAction("UserIndex");
+                    }
+                }
+
+                foreach (var error in result.Errors)
+                {
+                    ModelState.AddModelError("", error.Description);
+                }
+            }
+
+            return View(viewModel);
+        }
+
+        [HttpGet("Edit/{id}")]
+        public async Task<IActionResult> Edit(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var user = await userManager.FindByIdAsync(id.ToString());
+            if (user == null)
+            {
+                return NotFound();
+            }
+
+            var userRoles = await userRoleService.GetAllAsync();
+            var userRolesSelectList = userRoles.Select(x => new SelectListItem()
+            {
+                Text = x.Name,
+                Value = x.Id.ToString()
+            }).ToList();
+
+            var userStatuses = await userStatusService.GetAllAsync();
+            var userStatusesSelectList = userStatuses.Select(x => new SelectListItem()
+            {
+                Text = x.Name,
+                Value = x.Id.ToString()
+            }).ToList();
+
+            AdminUserEditViewModel viewModel = new AdminUserEditViewModel
+            {
+                FirstName = user.FirstName,
+                LastName = user.LastName,
+                Email = user.Email,
+                Age = user.Age,
+                Location = user.Location,
+                AvatarFilePath = user.AvatarFilePath,
+                AvatarImage = user.AvatarImage,
+                Description = user.Description,
+                UserRoles = userRolesSelectList,
+                UserStatuses = userStatusesSelectList
+            };
+
+            return View("~/Areas/Admin/Views/User/Edit.cshtml", viewModel);
+        }
+
+        [HttpPost("Edit/{id}")]
+        public async Task<IActionResult> Edit(AdminUserEditViewModel viewModel)
+        {
+            User user = await userManager.FindByIdAsync(viewModel.ID.ToString());
+            if (user == null)
+            {
+                return NotFound();
+            }
+
+            user.FirstName = viewModel.FirstName;
+            user.LastName = viewModel.LastName;
+            user.Email = viewModel.Email;
+            user.Age = viewModel.Age;
+            user.Location = viewModel.Location;
+            user.AvatarFilePath = viewModel.AvatarFilePath;
+            user.AvatarImage = viewModel.AvatarImage;
+            user.Description = viewModel.Description;
+            user.UserStatusID = viewModel.UserStatusID;
+            user.UserRoleID = viewModel.UserRoleID;
+
+            userService.Update(user);
+            return RedirectToAction("UserIndex", user);
+        }
+
+        [HttpGet("Delete/{id}")]
+        public async Task<IActionResult> Delete(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var user = await userManager.FindByIdAsync(id.ToString());
+            if (user == null)
+            {
+                return NotFound();
+            }
+
+            return View("~/Areas/Admin/Views/User/Delete.cshtml", user);
+        }
+
+        [HttpPost("Delete/{id}")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> DeleteConfirmed(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var user = await userManager.FindByIdAsync(id.ToString());
+            if (user != null)
+            {
+                var result = await userManager.DeleteAsync(user);
+                if (result.Succeeded)
+                {
+                    return RedirectToAction("UserIndex");
+                }
+
+                foreach (var error in result.Errors)
+                {
+                    ModelState.AddModelError("", error.Description);
+                }
+            }
+
+            return RedirectToAction("UserIndex", userManager.Users);
+        }
+
+        [HttpGet]
+        [AllowAnonymous]
+        public IActionResult AccessDenied()
+        {
+            return View("~/Areas/Admin/Views/Shared/AccessDenied.cshtml");
+        }
+    }
+}
