+++ GameSource/Areas/Admin/Views/User/Delete.cshtml
@@ -68,9 +68,9 @@
         </dd>
     </dl>
 
-    <form asp-area="Admin" asp-controller="Admin" asp-action="DeleteUser">
+    <form asp-area="Admin" asp-controller="User" asp-action="Delete">
         <input type="hidden" asp-for="Id" />
         <input type="submit" value="Delete" class="btn btn-danger" /> |
-        <a asp-area="Admin" asp-controller="Admin" asp-action="UserIndex">Back to List</a>
+        <a asp-area="Admin" asp-controller="User" asp-action="Index">Back to List</a>
     </form>
 </div>
