+++ GameSource/Areas/Admin/Views/User/Details.cshtml
@@ -80,6 +80,6 @@
     </dl>
 </div>
 <div>
-    <a asp-area="Admin" asp-controller="Admin" asp-action="EditUser" asp-route-id="@Model.User.Id">Edit</a> |
-    <a asp-area="Admin" asp-controller="Admin" asp-action="UserIndex">Back to List</a>
+    <a asp-area="Admin" asp-controller="User" asp-action="Edit" asp-route-id="@Model.User.Id">Edit</a> |
+    <a asp-area="Admin" asp-controller="User" asp-action="Index">Back to List</a>
 </div>
