+++ GameSource/Areas/Admin/Views/User/Edit.cshtml
@@ -11,7 +11,7 @@
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form asp-area="Admin" asp-controller="Admin" asp-action="EditUser">
+        <form asp-area="Admin" asp-controller="User" asp-action="Edit">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <input type="hidden" asp-for="ID" />
             <div class="form-group">
@@ -67,6 +67,6 @@
 </div>
 
 <div>
-    <a asp-area="Admin" asp-controller="Admin" asp-action="UserIndex">Back to List</a>
+    <a asp-area="Admin" asp-controller="User" asp-action="Index">Back to List</a>
 </div>
