+++ GameSource/Areas/Admin/Views/User/Register.cshtml
@@ -9,7 +9,7 @@
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form asp-area="Admin" asp-controller="Admin" asp-action="RegisterUser" method="post">
+        <form asp-area="Admin" asp-controller="User" asp-action="Register" method="post">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <div class="form-group">
                 <label asp-for="FirstName" class="control-label"></label>
@@ -49,6 +49,6 @@
 </div>
 
 <div>
-    <a asp-area="Admin" asp-controller="Admin" asp-action="UserIndex">Back to List</a>
+    <a asp-area="Admin" asp-controller="User" asp-action="Index">Back to List</a>
 </div>
