+++ GameSource/Areas/Admin/Views/UserRole/Edit.cshtml
@@ -11,7 +11,7 @@
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form asp-area="Admin" asp-controller="Admin" asp-action="EditUserRole">
+        <form asp-area="Admin" asp-controller="UserRole" asp-action="Edit">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <input type="hidden" asp-for="ID" />
             <div class="form-group">
@@ -32,6 +32,6 @@
 </div>
 
 <div>
-    <a asp-area="Admin" asp-controller="Admin" asp-action="UserRoleIndex">Back to List</a>
+    <a asp-area="Admin" asp-controller="UserRole" asp-action="Index">Back to List</a>
 </div>
