+++ GameSource/Areas/Admin/Views/UserStatus/Details.cshtml
@@ -20,6 +20,6 @@
     </dl>
 </div>
 <div>
-    <a asp-area="Admin" asp-controller="Admin" asp-action="EditUserStatus" asp-route-id="@Model.UserStatus.Id">Edit</a> |
-    <a asp-area="Admin" asp-controller="Admin" asp-action="UserStatusIndex">Back to List</a>
+    <a asp-area="Admin" asp-controller="UserStatus" asp-action="Edit" asp-route-id="@Model.UserStatus.Id">Edit</a> |
+    <a asp-area="Admin" asp-controller="UserStatus" asp-action="Index">Back to List</a>
 </div>
