+++ GameSource/Areas/Admin/Views/UserStatus/Edit.cshtml
@@ -11,7 +11,7 @@
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form asp-area="Admin" asp-controller="Admin" asp-action="EditUserStatus">
+        <form asp-area="Admin" asp-controller="UserStatus" asp-action="Edit">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <input type="hidden" asp-for="ID" />
             <div class="form-group">
@@ -27,6 +27,6 @@
 </div>
 
 <div>
-    <a asp-area="Admin" asp-controller="Admin" asp-action="UserStatusIndex">Back to List</a>
+    <a asp-area="Admin" asp-controller="UserStatus" asp-action="Index">Back to List</a>
 </div>
