+++ GameSource/Areas/Admin/Views/UserStatus/Index.cshtml
@@ -5,11 +5,11 @@
     Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
-<h1>Index</h1>
+<h1 class="text-center">Index</h1>
 
 <div class="container">
     <p>
-        <a class="btn btn-primary" asp-area="Admin" asp-controller="Admin" asp-action="CreateUserStatus">Create a new User Status</a>
+        <a class="btn btn-primary" asp-area="Admin" asp-controller="UserStatus" asp-action="Create">Create a new User Status</a>
     </p>
 
     <table class="table">
@@ -17,7 +17,6 @@
             <tr>
                 <th>ID</th>
                 <th>Name</th>
-                <th>Description</th>
                 <th>Actions</th>
             </tr>
         </thead>
@@ -28,9 +27,9 @@
                     <td>@Html.DisplayFor(modelItem => item.Id)</td>
                     <td>@Html.DisplayFor(modelItem => item.Name)</td>
                     <td>
-                        <a class="btn btn-primary" asp-area="Admin" asp-controller="Admin" asp-action="EditUserStatus" asp-route-id="@item.Id">Edit</a>
-                        <a class="btn btn-primary" asp-area="Admin" asp-controller="Admin" asp-action="UserStatusDetails" asp-route-id="@item.Id">Details</a>
-                        <a class="btn btn-primary" asp-area="Admin" asp-controller="Admin" asp-action="DeleteUserStatus" asp-route-id="@item.Id">Delete</a>
+                        <a class="btn btn-primary" asp-area="Admin" asp-controller="UserStatus" asp-action="Edit" asp-route-id="@item.Id">Edit</a>
+                        <a class="btn btn-primary" asp-area="Admin" asp-controller="UserStatus" asp-action="Details" asp-route-id="@item.Id">Details</a>
+                        <a class="btn btn-primary" asp-area="Admin" asp-controller="UserStatus" asp-action="Delete" asp-route-id="@item.Id">Delete</a>
                     </td>
                 </tr>
             }
