+++ GameSource/Views/Shared/_Layout.cshtml
@@ -69,9 +69,9 @@
                             <li class="nav-item dropdown">
                                 <a class="nav-link text-dark dropdown-toggle" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Administration</a>
                                 <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-                                    <li><a class="nav-link text-dark" asp-area="Admin" asp-controller="Admin" asp-action="UserIndex">View All Users</a></li>
-                                    <li><a class="nav-link text-dark" asp-area="Admin" asp-controller="Admin" asp-action="UserRoleIndex">View All User Roles</a></li>
-                                    <li><a class="nav-link text-dark" asp-area="Admin" asp-controller="Admin" asp-action="UserStatusIndex">View All User Statuses</a></li>
+                                    <li><a class="nav-link text-dark" asp-area="Admin" asp-controller="User" asp-action="Index">View All Users</a></li>
+                                    <li><a class="nav-link text-dark" asp-area="Admin" asp-controller="UserRole" asp-action="Index">View All User Roles</a></li>
+                                    <li><a class="nav-link text-dark" asp-area="Admin" asp-controller="UserStatus" asp-action="Index">View All User Statuses</a></li>
                                 </ul>
                             </li>
